Predicting Bank Client's Financial Product subscription using Scikit Learn and XGBoost for imbalance dataset.
We will illustrates the Machine Learning classification using the Gradient Boosted Tree. Gradient Boosted Tree, is usually a better choice compare to the logistic regression and other techniques. We will use the real life data set which is highly imbalance i.e the number of positive sample is much less than the number of negative samples.

We will walk the user to the the following conceptual steps

 Data Set Description.
 Exploratory Analysis to understand the data.
 Use various preprocessing to clean and prepare the data.
 Use naive XGBoost to run the classification.
 Use cross validation to get the model.
 Plot, precision recall curve and ROC curve.

We will then tune it and use weighted positive samples to improve classification performance.

 We will also talk about the following advanced techniques.
 Oversampling of majority class and Undersampling of minority class.
 SMOTE algorithms.